pyproject.toml
[tool.poetry]
name = "ibid_reportgen"
version = "0.1.0"
description = ""
authors = ["apteryxlabs <matthew@apteryxlabs.com>"]

[tool.poetry.dependencies]
python = ">=3.8 <3.9"
numpy = "^1.20.3"
PyQt5-sip = "^12.9.0"
matplotlib = "^3.4.2"
tqdm = "^4.61.0"
pandas = "^1.2.4"
pytest = "^6.2.4"
dominate = "^2.6.0"
PyPDF2 = "^1.26.0"
lxml = "~=4.6.3"
html5lib = "^1.1"
dateparser = "^1.0.0"
pymongo = "^3.11.4"
dnspython = "^2.1.0"
boto3 = "^1.17.92"
awscli = "^1.19.92"
google-api-core = "^1.30.0"
google-api-python-client = "^2.8.0"
google-auth = "^1.31.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"
googleapis-common-protos = "^1.53.0"
PyQt5 = "~=5.15.4"
beautifulsoup4 = "^4.9.3"
Jinja2 = "^3.0.1"
google-cloud-storage = "^1.38.0"
arrow = "^1.1.0"
apteryx = "^1.0.2"

[tool.poetry.dev-dependencies]
Faker = "^8.6.0"
pip = "^21.1.2"
jupyter = "^1.0.0"
jupyterlab = "^3.0.16"
tika = "^1.24"
bump2version = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
